requirements.txt
torch>=2.2.0
torchvision>=0.17.0
Pillow>=10.0.0
matplotlib>=3.8.0
